requirements.txt
ipykernel==5.1.0
ipython==7.5.0
jupyter==1.0.0
matplotlib==3.1.0
pandas==0.24.2
pylint==2.3.1
tqdm==4.31.1
scipy==1.3.0
lxml==4.3.3
seaborn==0.9.0
scikit-learn==0.21.2
beautifulsoup4==4.7.1
waterfallcharts==3.8